Clean up some peer dependency warnings
We don't need to transitively list rxjs and zone as peer deps: Angular itself lists them as such. We've cleaned this up in ng2-redux too.

Since this can be used with a raw Redux store, do we need Ng2 Redux to be a peer dep?

Fixed a typo in the readme.